pyproject.toml
[project]
name = "gs-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn",
    "streamlit",
    "pypdf>=6.1.2",
    "chromadb>=1.2.1",
    "sentence-transformers>=5.1.1",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.27",
    "langchain-community>=0.3.31",
    "langchain-google-genai>=2.0.10",
    "pytesseract>=0.3.13",
    "pdf2image>=1.17.0",
    "pdfplumber>=0.11.7",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "langchain-huggingface>=0.3.1",
]
